pyproject.toml
[tool.poetry]
name = "fastapi-firestore-mongodb-compat-api"
version = "0.1.0"
description = ""
authors = ["chandanmaurya0 <mchandan@student.nitw.ac.in>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.14"
uvicorn = "^0.35.0"
motor = "^3.7.1"
python-dotenv = "^1.1.1"
pydantic = {extras = ["email"], version = "^2.11.7"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.5.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
